Basketball Court Counter README
This an Android project done under the supervision of Andela Learning Community 2.0 (ALC 2.0).  It is done for the learning benefit of beginners in Android development to learn a basic android programming skills through this simple project..
This project depicts a mobile version of a Basketball Score Counter Board. In this project, the user has the privilege to choose and select any basketball team of his/her choice, make or input all necessary scores of choice and also capable of viewing the total score summary of the match.
Furthermore, he/she also have a choice of sharing the final scores on the social media or friends through the "Share" button added. I hope up will find this app so interesting. For the APK file and the screen shots, check this link: https://goo.gl/qL2wUL